Added note about native Promise.all
After playing around with how native Promise implementations work with this library, I noticed that you can't call the native `Promise.all` if it isn't being invoked as a method of `Promise`. Even though a lot of people are still only relying on polyfills at the moment, I think this is important to note for anyone using this in an environment where native promises are supported.